function location replaced
